backend/pyproject.toml
[tool.poetry]
name = "workplace-tools-backend"
version = "0.1.0"
description = "FastAPI backend for the Workplace Tools project"
authors = ["Workplace Tools Team <dev@workplacetools.example>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic-settings = "^2.3.0"
python-dotenv = "^1.0.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
slowapi = "^0.1.9"
supabase = "^2.4.0"
asyncpg = "^0.29.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["."]
